refactor(domain): 가독성 향상을 위한 도메인 패키지 분리
