+++ src/main/java/lotto/domain/generator/LottoGenerator.java
@@ -1,0 +1,7 @@
+package lotto.domain.generator;
+
+import lotto.domain.Lotto;
+
+public interface LottoGenerator {
+    Lotto generate();
+}
