+++ src/main/java/lotto/domain/generator/RandomLottoGenerator.java
@@ -1,6 +1,7 @@
-package lotto.domain;
+package lotto.domain.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import lotto.domain.Lotto;
 
 import java.util.List;
