+++ src/main/java/lotto/domain/generator/StubLottoGenerator.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.generator;
+
+import lotto.domain.Lotto;
 
 import java.util.List;
